combining envaware and bpp since both relate to cloud config
