+++ src/main/java/com/revature/rms/gateway/config/EurekaInstanceConfigBeanPostProcessor.java
@@ -5,17 +5,32 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.cloud.netflix.eureka.EurekaInstanceConfigBean;
+import org.springframework.context.EnvironmentAware;
 import org.springframework.context.annotation.Profile;
+import org.springframework.core.env.Environment;
+import org.springframework.core.env.PropertiesPropertySource;
+import org.springframework.core.env.StandardEnvironment;
 import org.springframework.stereotype.Component;
+
+import java.net.InetAddress;
+import java.net.UnknownHostException;
+import java.util.Properties;
 
 @Component
 @Profile("dev")
-public class EurekaInstanceConfigBeanPostProcessor implements BeanPostProcessor {
+public class EurekaInstanceConfigBeanPostProcessor implements BeanPostProcessor, EnvironmentAware {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(EurekaInstanceConfigBeanPostProcessor.class);
 
-    @Value("${FARGATE_IP}")
     private String fargateIp;
+
+    {
+        try {
+            fargateIp = InetAddress.getLocalHost().getHostAddress();
+        } catch (UnknownHostException e) {
+            LOGGER.warn("Could not get the Fargate instance ip address.");
+        }
+    }
 
     @Override
     public Object postProcessAfterInitialization(Object bean, String beanName) {
@@ -27,6 +42,21 @@ public class EurekaInstanceConfigBeanPostProcessor implements BeanPostProcessor 
         }
 
         return bean;
+    }
+
+    @Override
+    public void setEnvironment(Environment environment) {
+        Properties props = new Properties();
+        try {
+            LOGGER.info("setting FARGATE_IP address to {}", InetAddress.getLocalHost().getHostAddress());
+            props.setProperty("FARGATE_IP", InetAddress.getLocalHost().getHostAddress());
+            PropertiesPropertySource propertySource = new PropertiesPropertySource("ECS", props);
+            if (environment instanceof StandardEnvironment) {
+                ((StandardEnvironment) environment).getPropertySources().addFirst(propertySource);
+            }
+        } catch (UnknownHostException e) {
+            LOGGER.warn("Could not get the Fargate instance ip address.");
+        }
     }
 
 }
